Removed long double

+++ MandelbrotExplorer.h
@@ -23,7 +23,7 @@ private:
 // methods
 public:
     // constructor
-    MandelbrotExplorer(int size_in, long double zoomscale_in, int maxiter_in, int iterincr_in, int checkperiod_in);
+    MandelbrotExplorer(int size_in, double zoomscale_in, int maxiter_in, int iterincr_in, int checkperiod_in);
 
     // life-cycle of the explorer
     void start();
@@ -36,8 +36,8 @@ private:
     void zoom(bool zoomin);
 
     // returns the color of a pixel in the fractal
-    sf::Color get_mb_pixel(const long double cre, const long double cim);
+    sf::Color get_mb_pixel(const double cre, const double cim);
 
     // checks if the given point is in a cardoid/bulb
-    static bool in_cardoid(const long double cre, const long double cim);
+    static bool in_cardoid(const double cre, const double cim);
 };
